+++ udata_front/__init__.py
@@ -2,5 +2,5 @@
 udata customizations for data.gouv.fr
 '''
 
-__version__ = '6.1.4'
+__version__ = '6.1.5.dev'
 __description__ = 'udata customizations for data.gouv.fr'