+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Current (in progress)
+
+- Nothing yet
+
 ## 6.1.4 (2025-03-10)
 
 - Change unavailable resource style [#670](https://github.com/datagouv/udata-front/pull/670)